"check_style_score -> 0 of 3 passed. style score: 0/1 "

+++ ADS-1-Assignments/m16/16.1 BST/Solution.java
@@ -8,7 +8,7 @@ public class Solution {
 			switch (tokens[0]) {
 			case "put":
 				Book key = new Book(tokens[1], tokens[2],
-				                    Integer.parseInt(tokens[3]));
+				                    Float.parseFloat(tokens[3]));
 				int value = Integer.parseInt(tokens[4]);
 				bst.put(key, value);
 				break;
